requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
psycopg2
python-dotenv